+++ src/paddlefx/variable_dispatch.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from .dispatcher import Dispatcher
+
+# iter
+Dispatcher.register(
+    iter,
+    ("VariableBase",),
+    lambda variable: variable.get_iter(),
+)
