+++ src/paddlefx/variables/__init__.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
 
-from .base import BuiltinVariable, CallableVariable, ObjectVariable, VariableBase
+from .base import ObjectVariable, VariableBase
+from .basic import ConstantVariable
+from .callable import CallableVariable
 
 __all__ = [
     "VariableBase",
     "CallableVariable",
     "ObjectVariable",
-    "BuiltinVariable",
+    "ConstantVariable",
 ]
